Feature/test general

+++ Source/Core/RealEstateApp.Core.Application/ViewsModel/Property/PropertyDetailPublicConsult.cs
@@ -6,7 +6,7 @@
         public required string Code { get; set; }
         public decimal Price { get; set; }
         public required string Description { get; set; }
-        public double Size { get; set; }
+        public decimal Size { get; set; }
         public int Bedrooms { get; set; }
         public int Bathrooms { get; set; }
         public required string TypePropertyName { get; set; }
